requirements_mac.txt
torch
torchvision
eva-decord
numpy
scipy